allow full window draw
