+++ android/Omnivore/app/src/main/AndroidManifest.xml
@@ -22,6 +22,7 @@
         <activity
             android:name=".MainActivity"
             android:exported="true"
+            android:windowSoftInputMode="adjustResize"
             android:theme="@style/Theme.Omnivore">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
