+++ android/Omnivore/app/src/main/java/app/omnivore/omnivore/MainActivity.kt
@@ -4,12 +4,11 @@ import android.os.Bundle
 import android.view.View
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
 import androidx.activity.viewModels
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowCompat
 import androidx.core.view.WindowInsetsCompat
@@ -24,61 +23,65 @@ import app.omnivore.omnivore.ui.settings.SettingsViewModel
 import app.omnivore.omnivore.ui.theme.OmnivoreTheme
 import com.pspdfkit.PSPDFKit
 import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 
+@OptIn(DelicateCoroutinesApi::class)
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
-  override fun onCreate(savedInstanceState: Bundle?) {
-    super.onCreate(savedInstanceState)
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
 
-    val loginViewModel: LoginViewModel by viewModels()
-    val libraryViewModel: LibraryViewModel by viewModels()
-    val settingsViewModel: SettingsViewModel by viewModels()
-    val searchViewModel: SearchViewModel by viewModels()
-    val labelsViewModel: LabelsViewModel by viewModels()
-    val saveViewModel: SaveViewModel by viewModels()
-    val editInfoViewModel: EditInfoViewModel by viewModels()
+        val loginViewModel: LoginViewModel by viewModels()
+        val libraryViewModel: LibraryViewModel by viewModels()
+        val settingsViewModel: SettingsViewModel by viewModels()
+        val searchViewModel: SearchViewModel by viewModels()
+        val labelsViewModel: LabelsViewModel by viewModels()
+        val saveViewModel: SaveViewModel by viewModels()
+        val editInfoViewModel: EditInfoViewModel by viewModels()
 
-    val context = this
+        val context = this
 
-    GlobalScope.launch(Dispatchers.IO) {
-      val licenseKey = getString(R.string.pspdfkit_license_key)
+        GlobalScope.launch(Dispatchers.IO) {
+            val licenseKey = getString(R.string.pspdfkit_license_key)
 
-      if (licenseKey.length > 30) {
-        PSPDFKit.initialize(context, licenseKey)
-      } else {
-        PSPDFKit.initialize(context, null)
-      }
-    }
-
-    setContent {
-      OmnivoreTheme {
-        Box(
-          modifier = Modifier
-            .fillMaxSize()
-            .background(color = Color.Black)
-        ) {
-          RootView(
-            loginViewModel,
-            searchViewModel,
-            libraryViewModel,
-            settingsViewModel,
-            labelsViewModel,
-            saveViewModel,
-            editInfoViewModel)
+            if (licenseKey.length > 30) {
+                PSPDFKit.initialize(context, licenseKey)
+            } else {
+                PSPDFKit.initialize(context, null)
+            }
         }
-      }
-    }
 
-    // animate the view up when keyboard appears
-    WindowCompat.setDecorFitsSystemWindows(window, false)
-    val rootView = findViewById<View>(android.R.id.content).rootView
-    ViewCompat.setOnApplyWindowInsetsListener(rootView) { _, insets ->
-      val imeHeight = insets.getInsets(WindowInsetsCompat.Type.ime()).bottom
-      rootView.setPadding(0, 0, 0, imeHeight)
-      insets
+        enableEdgeToEdge()
+
+        setContent {
+            OmnivoreTheme {
+                Box(
+                    modifier = Modifier
+                        .fillMaxSize()
+                ) {
+                    RootView(
+                        loginViewModel,
+                        searchViewModel,
+                        libraryViewModel,
+                        settingsViewModel,
+                        labelsViewModel,
+                        saveViewModel,
+                        editInfoViewModel
+                    )
+                }
+            }
+        }
+
+        // animate the view up when keyboard appears
+        WindowCompat.setDecorFitsSystemWindows(window, false)
+        val rootView = findViewById<View>(android.R.id.content).rootView
+        ViewCompat.setOnApplyWindowInsetsListener(rootView) { _, insets ->
+            val imeHeight = insets.getInsets(WindowInsetsCompat.Type.ime()).bottom
+            rootView.setPadding(0, 0, 0, imeHeight)
+            insets
+        }
     }
-  }
 }
