+++ android/Omnivore/app/src/main/java/app/omnivore/omnivore/ui/auth/WelcomeScreen.kt
@@ -4,7 +4,7 @@ import android.annotation.SuppressLint
 import android.content.Intent
 import android.net.Uri
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.clickable
+import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.text.ClickableText
 import androidx.compose.material3.*
@@ -14,7 +14,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.AnnotatedString
@@ -23,154 +22,173 @@ import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.unit.dp
 import app.omnivore.omnivore.R
 import app.omnivore.omnivore.ui.theme.OmnivoreTheme
+import com.google.accompanist.systemuicontroller.rememberSystemUiController
 import com.google.android.gms.common.GoogleApiAvailability
 import kotlinx.coroutines.launch
 
 @Composable
 fun WelcomeScreen(viewModel: LoginViewModel) {
-  OmnivoreTheme(darkTheme = false) {
-    Surface(modifier = Modifier.fillMaxSize(), color = Color(0xFFFCEBA8)) {
-      WelcomeScreenContent(viewModel = viewModel)
+
+    val systemUiController = rememberSystemUiController()
+    val useDarkIcons = !isSystemInDarkTheme()
+
+    DisposableEffect(systemUiController, useDarkIcons) {
+        systemUiController.setSystemBarsColor(
+            color = Color.Black,
+            darkIcons = true
+        )
+        onDispose {
+            systemUiController.setSystemBarsColor(
+                color = Color.Black,
+                darkIcons = useDarkIcons
+            )
+        }
     }
-  }
+
+    OmnivoreTheme(darkTheme = false) {
+        Surface(modifier = Modifier.fillMaxSize(), color = Color(0xFFFCEBA8)) {
+            WelcomeScreenContent(viewModel = viewModel)
+        }
+    }
 }
 
 @SuppressLint("CoroutineCreationDuringComposition")
 @Composable
 fun WelcomeScreenContent(viewModel: LoginViewModel) {
-  val registrationState: RegistrationState by viewModel
-    .registrationStateLiveData
-    .observeAsState(RegistrationState.SocialLogin)
+    val registrationState: RegistrationState by viewModel
+        .registrationStateLiveData
+        .observeAsState(RegistrationState.SocialLogin)
 
-  val snackBarHostState = remember { SnackbarHostState() }
-  val coroutineScope = rememberCoroutineScope()
-  val focusManager = LocalFocusManager.current
+    val snackBarHostState = remember { SnackbarHostState() }
+    val coroutineScope = rememberCoroutineScope()
 
-  Column(
-    verticalArrangement = Arrangement.SpaceAround,
-    horizontalAlignment = Alignment.Start,
-    modifier = Modifier
-      .fillMaxSize()
-      .padding(horizontal = 16.dp)
-      .clickable { focusManager.clearFocus() }
-  ) {
-    Spacer(modifier = Modifier.height(50.dp))
-    Image(
-      painter = painterResource(id = R.drawable.ic_omnivore_name_logo),
-      contentDescription = "Omnivore Icon with Name"
-    )
-    Spacer(modifier = Modifier.height(50.dp))
-
-    when(registrationState) {
-      RegistrationState.EmailSignIn -> {
-        EmailLoginView(viewModel = viewModel)
-      }
-      RegistrationState.EmailSignUp -> {
-        EmailSignUpView(viewModel = viewModel)
-      }
-      RegistrationState.SelfHosted -> {
-        SelfHostedView(viewModel = viewModel)
-      }
-      RegistrationState.SocialLogin -> {
-        Text(
-          text = stringResource(id = R.string.welcome_title),
-          style = MaterialTheme.typography.headlineLarge
+    Column(
+        verticalArrangement = Arrangement.SpaceAround,
+        horizontalAlignment = Alignment.Start,
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(16.dp)
+    ) {
+        Spacer(modifier = Modifier.height(50.dp))
+        Image(
+            painter = painterResource(id = R.drawable.ic_omnivore_name_logo),
+            contentDescription = "Omnivore Icon with Name"
         )
-
-        Text(
-          text = stringResource(id = R.string.welcome_subtitle),
-          style = MaterialTheme.typography.titleSmall
-        )
-
-        MoreInfoButton()
-
         Spacer(modifier = Modifier.height(50.dp))
 
-        AuthProviderView(viewModel = viewModel)
-      }
-      RegistrationState.PendingUser -> {
-        CreateUserProfileView(viewModel = viewModel)
-      }
+        when (registrationState) {
+            RegistrationState.EmailSignIn -> {
+                EmailLoginView(viewModel = viewModel)
+            }
+
+            RegistrationState.EmailSignUp -> {
+                EmailSignUpView(viewModel = viewModel)
+            }
+
+            RegistrationState.SelfHosted -> {
+                SelfHostedView(viewModel = viewModel)
+            }
+
+            RegistrationState.SocialLogin -> {
+                Text(
+                    text = stringResource(id = R.string.welcome_title),
+                    style = MaterialTheme.typography.headlineLarge
+                )
+
+                Text(
+                    text = stringResource(id = R.string.welcome_subtitle),
+                    style = MaterialTheme.typography.titleSmall
+                )
+
+                MoreInfoButton()
+
+                Spacer(modifier = Modifier.height(50.dp))
+
+                AuthProviderView(viewModel = viewModel)
+            }
+
+            RegistrationState.PendingUser -> {
+                CreateUserProfileView(viewModel = viewModel)
+            }
+        }
+
+        Spacer(modifier = Modifier.weight(1.0F))
     }
 
-    Spacer(modifier = Modifier.weight(1.0F))
-  }
+    if (viewModel.errorMessage != null) {
+        coroutineScope.launch {
+            val result = snackBarHostState
+                .showSnackbar(
+                    viewModel.errorMessage!!,
+                    actionLabel = "Dismiss",
+                    duration = SnackbarDuration.Indefinite
+                )
+            when (result) {
+                SnackbarResult.ActionPerformed -> viewModel.resetErrorMessage()
+                else -> {}
+            }
+        }
 
-  if (viewModel.errorMessage != null) {
-    coroutineScope.launch {
-      val result = snackBarHostState
-        .showSnackbar(
-          viewModel.errorMessage!!,
-          actionLabel = "Dismiss",
-          duration = SnackbarDuration.Indefinite
-        )
-      when (result) {
-        SnackbarResult.ActionPerformed -> viewModel.resetErrorMessage()
-        else -> {}
-      }
+        SnackbarHost(hostState = snackBarHostState)
     }
-
-    SnackbarHost(hostState = snackBarHostState)
-  }
 }
 
 @Composable
 fun AuthProviderView(viewModel: LoginViewModel) {
-  val isGoogleAuthAvailable: Boolean = GoogleApiAvailability
-    .getInstance()
-    .isGooglePlayServicesAvailable(LocalContext.current) == 0
+    val isGoogleAuthAvailable: Boolean = GoogleApiAvailability
+        .getInstance()
+        .isGooglePlayServicesAvailable(LocalContext.current) == 0
 
-  Row(
-    horizontalArrangement = Arrangement.Center
-  ) {
-    Spacer(modifier = Modifier.weight(1.0F))
-    Column(
-      verticalArrangement = Arrangement.spacedBy(8.dp),
-      horizontalAlignment = Alignment.CenterHorizontally
+    Row(
+        horizontalArrangement = Arrangement.Center
     ) {
-      if (isGoogleAuthAvailable) {
-        GoogleAuthButton(viewModel)
-      }
+        Spacer(modifier = Modifier.weight(1.0F))
+        Column(
+            verticalArrangement = Arrangement.spacedBy(8.dp),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            if (isGoogleAuthAvailable) {
+                GoogleAuthButton(viewModel)
+            }
 
-      AppleAuthButton(viewModel)
+            AppleAuthButton(viewModel)
 
-      ClickableText(
-        text = AnnotatedString(stringResource(R.string.welcome_screen_action_continue_with_email)),
-        style = MaterialTheme.typography.titleMedium
-          .plus(TextStyle(textDecoration = TextDecoration.Underline)),
-        onClick = { viewModel.showEmailSignIn() }
-      )
+            ClickableText(
+                text = AnnotatedString(stringResource(R.string.welcome_screen_action_continue_with_email)),
+                style = MaterialTheme.typography.titleMedium
+                    .plus(TextStyle(textDecoration = TextDecoration.Underline)),
+                onClick = { viewModel.showEmailSignIn() }
+            )
 
-      Spacer(modifier = Modifier.weight(1.0F))
+            Spacer(modifier = Modifier.weight(1.0F))
 
-      ClickableText(
-        text = AnnotatedString(stringResource(R.string.welcome_screen_action_self_hosting_options)),
-        style = MaterialTheme.typography.titleMedium
-          .plus(TextStyle(textDecoration = TextDecoration.Underline)),
-        onClick = { viewModel.showSelfHostedSettings() },
-        modifier = Modifier
-          .padding(vertical = 10.dp)
-      )
+            ClickableText(
+                text = AnnotatedString(stringResource(R.string.welcome_screen_action_self_hosting_options)),
+                style = MaterialTheme.typography.titleMedium
+                    .plus(TextStyle(textDecoration = TextDecoration.Underline)),
+                onClick = { viewModel.showSelfHostedSettings() },
+                modifier = Modifier
+                    .padding(vertical = 10.dp)
+            )
+        }
+        Spacer(modifier = Modifier.weight(1.0F))
     }
-    Spacer(modifier = Modifier.weight(1.0F))
-  }
 }
 
 @Composable
 fun MoreInfoButton() {
-  val context = LocalContext.current
-  val intent = remember { Intent(Intent.ACTION_VIEW, Uri.parse("https://omnivore.app/about")) }
+    val context = LocalContext.current
+    val intent = remember { Intent(Intent.ACTION_VIEW, Uri.parse("https://omnivore.app/about")) }
 
-  ClickableText(
-    text = AnnotatedString(
-      stringResource(id = R.string.learn_more),
-    ),
-    style = MaterialTheme.typography.titleSmall
-      .plus(TextStyle(textDecoration = TextDecoration.Underline)),
-    onClick = {
-      context.startActivity(intent)
-    },
-    modifier = Modifier.padding(vertical = 6.dp)
-  )
+    ClickableText(
+        text = AnnotatedString(
+            stringResource(id = R.string.learn_more),
+        ),
+        style = MaterialTheme.typography.titleSmall
+            .plus(TextStyle(textDecoration = TextDecoration.Underline)),
+        onClick = {
+            context.startActivity(intent)
+        },
+        modifier = Modifier.padding(vertical = 6.dp)
+    )
 }
-
