+++ android/Omnivore/app/src/main/java/app/omnivore/omnivore/ui/root/RootView.kt
@@ -1,14 +1,10 @@
 package app.omnivore.omnivore.ui.root
 
-import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.systemBarsPadding
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
@@ -18,109 +14,98 @@ import app.omnivore.omnivore.ui.auth.WelcomeScreen
 import app.omnivore.omnivore.ui.components.LabelsViewModel
 import app.omnivore.omnivore.ui.editinfo.EditInfoViewModel
 import app.omnivore.omnivore.ui.library.LibraryView
-import app.omnivore.omnivore.ui.library.SearchView
 import app.omnivore.omnivore.ui.library.LibraryViewModel
+import app.omnivore.omnivore.ui.library.SearchView
 import app.omnivore.omnivore.ui.library.SearchViewModel
 import app.omnivore.omnivore.ui.save.SaveViewModel
 import app.omnivore.omnivore.ui.settings.PolicyWebView
 import app.omnivore.omnivore.ui.settings.SettingsView
 import app.omnivore.omnivore.ui.settings.SettingsViewModel
-import com.google.accompanist.systemuicontroller.rememberSystemUiController
 
 @Composable
 fun RootView(
-  loginViewModel: LoginViewModel,
-  searchViewModel: SearchViewModel,
-  libraryViewModel: LibraryViewModel,
-  settingsViewModel: SettingsViewModel,
-  labelsViewModel: LabelsViewModel,
-  saveViewModel: SaveViewModel,
-  editInfoViewModel: EditInfoViewModel,
+    loginViewModel: LoginViewModel,
+    searchViewModel: SearchViewModel,
+    libraryViewModel: LibraryViewModel,
+    settingsViewModel: SettingsViewModel,
+    labelsViewModel: LabelsViewModel,
+    saveViewModel: SaveViewModel,
+    editInfoViewModel: EditInfoViewModel,
 ) {
-  val hasAuthToken: Boolean by loginViewModel.hasAuthTokenLiveData.observeAsState(false)
-  val systemUiController = rememberSystemUiController()
-  val useDarkIcons = !isSystemInDarkTheme()
+    val hasAuthToken: Boolean by loginViewModel.hasAuthTokenLiveData.observeAsState(false)
 
-  DisposableEffect(systemUiController, useDarkIcons) {
-    systemUiController.setSystemBarsColor(
-      color = Color.Black,
-      darkIcons = false
-    )
+    Box {
+        if (hasAuthToken) {
+            PrimaryNavigator(
+                loginViewModel = loginViewModel,
+                searchViewModel = searchViewModel,
+                libraryViewModel = libraryViewModel,
+                settingsViewModel = settingsViewModel,
+                labelsViewModel = labelsViewModel,
+                saveViewModel = saveViewModel,
+                editInfoViewModel = editInfoViewModel,
+            )
+        } else {
+            WelcomeScreen(viewModel = loginViewModel)
+        }
 
-    onDispose {}
-  }
-
-  Box(
-    modifier = Modifier
-      .systemBarsPadding()
-  ) {
-    if (hasAuthToken) {
-      PrimaryNavigator(
-        loginViewModel = loginViewModel,
-        searchViewModel = searchViewModel,
-        libraryViewModel = libraryViewModel,
-        settingsViewModel = settingsViewModel,
-        labelsViewModel = labelsViewModel,
-        saveViewModel = saveViewModel,
-        editInfoViewModel = editInfoViewModel,
-      )
-    } else {
-      WelcomeScreen(viewModel = loginViewModel)
+        DisposableEffect(hasAuthToken) {
+            if (hasAuthToken) {
+                loginViewModel.registerUser()
+            }
+            onDispose {}
+        }
     }
-
-    DisposableEffect(hasAuthToken) {
-      if (hasAuthToken) {
-        loginViewModel.registerUser()
-      }
-      onDispose {}
-    }
-  }
 }
 
 @Composable
 fun PrimaryNavigator(
-  loginViewModel: LoginViewModel,
-  libraryViewModel: LibraryViewModel,
-  searchViewModel: SearchViewModel,
-  settingsViewModel: SettingsViewModel,
-  labelsViewModel: LabelsViewModel,
-  saveViewModel: SaveViewModel,
-  editInfoViewModel: EditInfoViewModel,
+    loginViewModel: LoginViewModel,
+    libraryViewModel: LibraryViewModel,
+    searchViewModel: SearchViewModel,
+    settingsViewModel: SettingsViewModel,
+    labelsViewModel: LabelsViewModel,
+    saveViewModel: SaveViewModel,
+    editInfoViewModel: EditInfoViewModel,
 ) {
-  val navController = rememberNavController()
+    val navController = rememberNavController()
 
-  NavHost(navController = navController, startDestination = Routes.Library.route) {
-    composable(Routes.Library.route) {
-      LibraryView(
-        libraryViewModel = libraryViewModel,
-        navController = navController,
-        labelsViewModel = labelsViewModel,
-        saveViewModel = saveViewModel,
-        editInfoViewModel = editInfoViewModel,
-      )
-    }
+    NavHost(navController = navController, startDestination = Routes.Library.route) {
+        composable(Routes.Library.route) {
+            LibraryView(
+                libraryViewModel = libraryViewModel,
+                navController = navController,
+                labelsViewModel = labelsViewModel,
+                saveViewModel = saveViewModel,
+                editInfoViewModel = editInfoViewModel,
+            )
+        }
 
-    composable(Routes.Search.route) {
-      SearchView(
-        viewModel = searchViewModel,
-        navController = navController
-      )
-    }
+        composable(Routes.Search.route) {
+            SearchView(
+                viewModel = searchViewModel,
+                navController = navController
+            )
+        }
 
-    composable(Routes.Settings.route) {
-      SettingsView(loginViewModel = loginViewModel, settingsViewModel = settingsViewModel, navController = navController)
-    }
+        composable(Routes.Settings.route) {
+            SettingsView(
+                loginViewModel = loginViewModel,
+                settingsViewModel = settingsViewModel,
+                navController = navController
+            )
+        }
 
-    composable(Routes.Documentation.route) {
-      PolicyWebView(navController = navController, url = "https://docs.omnivore.app")
-    }
+        composable(Routes.Documentation.route) {
+            PolicyWebView(navController = navController, url = "https://docs.omnivore.app")
+        }
 
-    composable(Routes.PrivacyPolicy.route) {
-      PolicyWebView(navController = navController, url = "https://omnivore.app/privacy")
-    }
+        composable(Routes.PrivacyPolicy.route) {
+            PolicyWebView(navController = navController, url = "https://omnivore.app/privacy")
+        }
 
-    composable(Routes.TermsAndConditions.route) {
-      PolicyWebView(navController = navController, url = "https://omnivore.app/app/terms")
+        composable(Routes.TermsAndConditions.route) {
+            PolicyWebView(navController = navController, url = "https://omnivore.app/app/terms")
+        }
     }
-  }
 }
